pyproject.toml
[project]
name = "astra"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.13.1",
    "fastapi>=0.125.0",
    "langchain>=1.2.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.5",
    "numpy>=2.3.5",
    "openai>=2.13.0",
    "pymupdf>=1.26.7",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "sentence-transformers>=5.2.0",
    "uvicorn>=0.38.0",
]
